🚚 [projects] Move function `loadtemplate` from `services` to `projects`
