+++ src/cutty/entrypoints/cli/cookiecutter.py
@@ -7,7 +7,7 @@ import click
 
 from cutty.entrypoints.cli.create import extra_context_callback
 from cutty.projects.generate import generate
-from cutty.services.loadtemplate import loadtemplate
+from cutty.projects.loadtemplate import loadtemplate
 from cutty.templates.domain.bindings import Binding
