+++ src/cutty/entrypoints/cli/errors.py
@@ -2,6 +2,7 @@
 import pathlib
 from typing import NoReturn
 
+from cutty.projects.generate import EmptyTemplateError
 from cutty.repositories.adapters.fetchers.file import FileFetcherError
 from cutty.repositories.adapters.fetchers.git import GitFetcherError
 from cutty.repositories.adapters.fetchers.http import HTTPFetcherError
@@ -10,7 +11,6 @@ from cutty.repositories.adapters.fetchers.mercurial import HgNotFoundError
 from cutty.repositories.adapters.providers.git import RevisionNotFoundError
 from cutty.repositories.domain.mounters import UnsupportedRevisionError
 from cutty.repositories.domain.registry import UnknownLocationError
-from cutty.services.generate import EmptyTemplateError
 from cutty.services.link import TemplateNotSpecifiedError
 from cutty.util.exceptionhandlers import exceptionhandler
