+++ src/cutty/projects/create.py
@@ -5,7 +5,7 @@ import pygit2
 
 from cutty.projects.common import createcommitmessage
 from cutty.projects.common import LATEST_BRANCH
-from cutty.services.loadtemplate import TemplateMetadata
+from cutty.projects.loadtemplate import TemplateMetadata
 from cutty.util import git
