+++ src/cutty/services/link.py
@@ -7,7 +7,7 @@ from typing import Optional
 from cutty.errors import CuttyError
 from cutty.projects.generate import generate
 from cutty.projects.link import linkproject
-from cutty.services.loadtemplate import loadtemplate
+from cutty.projects.loadtemplate import loadtemplate
 from cutty.templates.adapters.cookiecutter.projectconfig import readcookiecutterjson
 from cutty.templates.domain.bindings import Binding
 from cutty.util.git import Repository
